+++ lib/string/index.ts
@@ -1,0 +1,22 @@
+'use strict';
+
+import {humanizeBytes}      from './humanizeBytes';
+import {humanizeNumber}     from './humanizeNumber';
+import {isString}           from './is';
+import {isNotEmptyString}   from './isNotEmpty';
+import {isStringBetween}    from './isBetween';
+import {shorten}            from './shorten';
+
+export {
+    humanizeBytes,
+    humanizeNumber,
+    isString,
+    isString as is,
+    isNotEmptyString,
+    isNotEmptyString as isNotEmpty,
+    isNotEmptyString as isNeString,
+    isNotEmptyString as isNe,
+    isStringBetween,
+    isStringBetween as isBetween,
+    shorten,
+};